+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/ops/package.scala
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2016 The BigDL Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.intel.analytics.bigdl.nn
+
+import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
+import com.intel.analytics.bigdl.utils.Table
+
+import scala.reflect.ClassTag
+
+package object ops {
+  object Add {
+    def apply[T: ClassTag]()(implicit ev: TensorNumeric[T]): Operation[Table, T]
+    = ModuleToOperation[Table, T](CAddTable())
+  }
+
+  object Subtract {
+    def apply[T: ClassTag]()(implicit ev: TensorNumeric[T]): Operation[Table, T]
+    = ModuleToOperation[Table, T](CSubTable())
+  }
+
+  object Multiply {
+    def apply[T: ClassTag]()(implicit ev: TensorNumeric[T]): Operation[Table, T]
+    = ModuleToOperation[Table, T](CMulTable())
+  }
+
+  object Divide {
+    def apply[T: ClassTag]()(implicit ev: TensorNumeric[T]): Operation[Table, T]
+    = ModuleToOperation[Table, T](CDivTable())
+  }
+}
